Refactor models into separate files for better maintainability
